dsf_backend_adk/pyproject.toml
[project]
name = "dsf-backend-adk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.6.1",
    "jinja2>=3.1.6",
    "pandas>=2.3.1",
]
